aaa (+2 squashed commit)
Squashed commit:

[2b9fce4] bbb

[a8bd61b] # This is a combination of 2 commits.

Colorbox

bbb
