+++ .gitignore
@@ -14,3 +14,4 @@ coverage.lcov
 .nyc_output
 yarn-error.log
 tmp
+*.lock
