+++ credentials.json
@@ -1,0 +1,25 @@
+{
+  "flickr": {
+    "api_key": "",
+    "app_secret": ""
+  },
+  "twitter": {
+    "consumer_key": "",
+    "consumer_secret": "",
+    "access_token": "",
+    "access_token_secret": ""
+  },
+  "facebook": {
+    "app_id": "",
+    "app_secret": ""
+  },
+  "instagram": {
+    "client_id": "53cbd338e38643ba96179cdb50a333a3",
+    "client_secret": "8225835440eb4689b1632b808d5ccea7"
+  },
+  "google": {
+    "maps": {
+      "api_key": ""
+    }
+  }
+}
