+++ index.js
@@ -8,6 +8,7 @@ const lout = require('lout');
 const notifier = require('node-notifier');
 const inert = require('inert');
 const vision = require('vision');
+const instagram = require('instagram-node');
 
 require('tuxharness');
 
@@ -23,6 +24,8 @@ const libResize = require('./plugins/resize/lib/index');
 const libVideo = require('./plugins/video/lib/index');
 const logMod = require('./plugins/log/lib/log');
 const pkg = require('./package');
+
+const credentials = require('./credentials.json');
 
 require('babel-core/register')({
   presets: ['react', 'es2015'],
@@ -70,4 +73,68 @@ server.views({
   path: './',
   partialsPath: './',
   relativeTo: __dirname,
+});
+
+const ig = instagram.instagram();
+const redirectLandingAddress = 'http://localhost:8080/api/instagram-login';
+
+server.route({
+  method: 'GET',
+  path: '/api/instagram-login',
+  handler: (request, reply) => {
+    if (request.query.code) {
+      ig.authorize_user(request.query.code, redirectLandingAddress, (authError, result) => {
+        if (authError) {
+          reply(`Error found ${authError.message}`);
+          return;
+        }
+
+        credentials.instagram.access_token = result.access_token;
+
+        ig.use({
+          access_token: credentials.instagram.access_token,
+          client_secret: credentials.instagram.client_secret,
+        });
+
+        // error, medias, pagination, remaining, limit
+        ig.tag_media_recent('vancouver', { count: 10 }, (mediaError, media) => {
+          if (mediaError) {
+            reply(`Error found ${mediaError.message}`);
+            return;
+          }
+
+          reply(media);
+        });
+      });
+    } else {
+      ig.use({
+        client_id: credentials.instagram.client_id,
+        client_secret: credentials.instagram.client_secret,
+      });
+
+      reply()
+        .redirect(ig.get_authorization_url(redirectLandingAddress, { scope: ['public_content'] }));
+    }
+  },
+});
+
+server.route({
+  method: 'GET',
+  path: '/api/instagram',
+  handler: (request, reply) => {
+    ig.use({
+      access_token: credentials.instagram.access_token,
+      client_secret: credentials.instagram.client_secret,
+    });
+
+    // error, medias, pagination, remaining, limit
+    ig.tag_media_recent('vancouver', { count: 10 }, (mediaError, media) => {
+      if (mediaError) {
+        reply(`Error found ${mediaError.message}`);
+        return;
+      }
+
+      reply(media);
+    });
+  },
 });
