+++ package.json
@@ -41,7 +41,7 @@
     "inert": "^4.0.2",
     "instagram-node": "^0.5.8",
     "joi": "^10.0.0",
-    "justified-gallery": "^0.1.0",
+    "jquery-colorbox": "^1.6.4",
     "lout": "^9.2.0",
     "mime-types": "^2.1.12",
     "mkdirp": "^0.5.0",
